1) Create new event page now works. A new event can be created and the event is added to the Events collection in the mongo DB. The host of the event is set to the currently logged in user (i.e. the objectID of the user's entry in the Users collection). After event creation, all input fields are cleared and the user stays on the same page. We can add code to take the user to another page on event creation.
2) Removed the host name field from the create event page. The host is the currently logged in user.
Also cleaned up the other fields on the page to match the Event model (see next item).

3) The Event model is cleaned up. No more keyword #2 or #3. Just a keywords field which is a string.
Added fields for event date and time. Also made the event name unique.

4) createEvent function added to client/src/utils/API.js

5) Fixed an issue in App.js on client side that was causing an infinite loop of an api call.
Also added a comment explaining what to change if you want to disable the login function in your local environment (see line 25).
All that is needed is to change a false to true. We don't need to comment and uncomment a bunch of lines of code.
Please don't disable the login function in the master branch.

6) Updated the create event code in the events controller.

7) Fixed the routes in routes/api/events.

8) Changed "Login" button to "Login with LinkedIn"